==> Notebooks/0. Merging Datasets.ipynb <==
# Mount Drive


# Load library google drive
from google.colab import drive
drive.mount('gdrive')

root = '/content/gdrive/MyDrive/Colab Notebooks/ML Bangkit/'

import pandas as pd
import numpy as np

datasets = pd.DataFrame(columns=['text', 'label'])

# Load Hotel Dataset

hotel_dataset_path = 'Datasets/classes/hotel/hotel_review.csv'

df = pd.read_csv(root + hotel_dataset_path, nrows=1000, encoding='ISO-8859-1')
df.head()

for text in df.Review:
  datasets.loc[len(datasets.index)] = [text, 'hotel']

# Load Restaurant Dataset

restaurant_dataset_path = 'Datasets/classes/restaurant/Restaurant_Reviews.tsv'

df = pd.read_csv(root + restaurant_dataset_path, sep='\t')
df.head()

for text in df.Review:
  if text.strip() == '':
    print('ada')
    continue
  datasets.loc[len(datasets.index)] = [text, 'restaurant']

# Load Museum Dataset

museum_dataset_path = 'Datasets/classes/museum/tripadvisor_merged.csv'

df = pd.read_csv(root + museum_dataset_path)
df = df.dropna(subset=['Description'])
df.head()

for text in df.Description:
  datasets.loc[len(datasets.index)] = [text, 'museum']

# Load Temple Dataset

temple_dataset_path = 'Datasets/classes/temple/temple_reviews.csv'

df = pd.read_csv(root + temple_dataset_path)
df.head()

for text in df.caption:
  datasets.loc[len(datasets.index)] = [text, 'temple']

# Load Beach Dataset

beach_dataset_path = 'Datasets/classes/beach/beach_reviews.csv'

df = pd.read_csv(root + beach_dataset_path)
df.head()

for text in df.caption:
  datasets.loc[len(datasets.index)] = [text, 'beach']

# Load Food Dataset

food_dataset_path1 = 'Datasets/classes/food/food1.xlsx'
food_dataset_path2 = ''

# Save Merged to Drive

datasets = datasets.reset_index(drop=True)

datasets.to_excel(root + 'Datasets/FIXED/fixed_data.xlsx', index=False)
