==> Notebooks/1. Preprocessing.ipynb <==
# Mount Google Drive

# Load library google drive
from google.colab import drive
drive.mount('gdrive')

root = '/content/gdrive/MyDrive/Colab Notebooks/ML Bangkit/'

# Load Dataset

import pandas as pd
import numpy as np

df = pd.read_excel(root + 'Datasets/FIXED/fixed_data.xlsx')
df.reset_index(drop=True)
df.head()

# Declare Preprocessing Method

import re
import string
import nltk
from nltk.stem import PorterStemmer
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize

# Lowering Text
def lower(text):
  return text.lower()

# Lexical Analysis
exclist = string.punctuation + string.digits
def remove_punc_and_number(text):
  newtext = ''.join(x for x in text if x not in exclist)
  return newtext

# Remove Leading Space
def remove_space(text):
  splited = text.split()
  splited = [word.strip() for word in splited]
  return ' '.join(splited)

# Stemming
ps = PorterStemmer()
nltk.download('punkt')
def stem(text):
  words = word_tokenize(text)
  stemmed = [ps.stem(word) for word in words]
  return ' '.join(stemmed)

# Lemmatizing
lemmatizer = WordNetLemmatizer()
nltk.download('wordnet')
nltk.download('omw-1.4')
def lemmatize(text):
  words = word_tokenize(text)
  lemmatized = [lemmatizer.lemmatize(word) for word in words]
  return ' '.join(lemmatized)

# Remove Stopwords
nltk.download('stopwords')
stopwords = nltk.corpus.stopwords.words('english')
def remove_stopwords(text):
  words = word_tokenize(text)
  words = [word for word in words if word not in stopwords]
  return ' '.join(words)

# Preprocessing Umum
Lower, Remove Punctuation and Number, remove leading space, Remove Stopwords

df.text = [lower(str(text)) for text in df.text]
df.text = [remove_punc_and_number(text) for text in df.text]
df.text = [remove_space(text) for text in df.text]
df.text = [remove_stopwords(text) for text in df.text]

# Preprocessing Skenario 1
With Stem

df1 = df.copy()
df1.text = [stem(text) for text in df1.text]
df1.to_excel(root + 'Datasets/preprocessed/data_skenario_1.xlsx')

# Preprocessing Skenario 2
With Lemmatize

df2 = df.copy()
df2.text = [lemmatize(text) for text in df2.text]
df2.to_excel(root + 'Datasets/preprocessed/data_skenario_2.xlsx')

# Preprocessing Skenario 3
W/o Stem and Lemmatize

df.to_excel(root + 'Datasets/preprocessed/data_skenario_3.xlsx')
