==> Notebooks/2. Splitting Dataset.ipynb <==
# Mount Google Drive

# Load library google drive
from google.colab import drive
drive.mount('gdrive')

root = '/content/gdrive/MyDrive/Colab Notebooks/ML Bangkit/'

# Load Dataset

import pandas as pd
import numpy as np

df1 = pd.read_excel(root + 'Datasets/preprocessed/data_skenario_1.xlsx')
df1.reset_index(drop=True)
df1 = df1[['text', 'label']]

df2 = pd.read_excel(root + 'Datasets/preprocessed/data_skenario_2.xlsx')
df2.reset_index(drop=True)
df2 = df2[['text', 'label']]

df3 = pd.read_excel(root + 'Datasets/preprocessed/data_skenario_3.xlsx')
df3.reset_index(drop=True)
df3 = df3[['text', 'label']]

# Split Dataset

from sklearn.model_selection import train_test_split

train, test = train_test_split(df1, test_size=0.2, random_state=42)
train, val = train_test_split(train, test_size=0.25, random_state=40)
train.to_excel(root + 'Datasets/split/skenario 1/train.xlsx')
test.to_excel(root + 'Datasets/split/skenario 1/test.xlsx')
val.to_excel(root + 'Datasets/split/skenario 1/val.xlsx')

train, test = train_test_split(df2, test_size=0.2, random_state=42)
train, val = train_test_split(train, test_size=0.25, random_state=40)
train.to_excel(root + 'Datasets/split/skenario 2/train.xlsx')
test.to_excel(root + 'Datasets/split/skenario 2/test.xlsx')
val.to_excel(root + 'Datasets/split/skenario 2/val.xlsx')

train, test = train_test_split(df3, test_size=0.2, random_state=42)
train, val = train_test_split(train, test_size=0.25, random_state=40)
train.to_excel(root + 'Datasets/split/skenario 3/train.xlsx')
test.to_excel(root + 'Datasets/split/skenario 3/test.xlsx')
val.to_excel(root + 'Datasets/split/skenario 3/val.xlsx')
