==> Notebooks/4. Testing.ipynb <==
# Mount Google Drive

# Load library google drive
from google.colab import drive
drive.mount('gdrive')

root = '/content/gdrive/MyDrive/Colab Notebooks/ML Bangkit/'

# Load Testing Data

import pandas as pd
import numpy as np

df = pd.read_excel(root + 'Datasets/split/skenario 3/val.xlsx')
df = df[['text', 'label']]
df.head()

# Load Tokenizer

import pickle

f = open(root + 'Saved/tokenizer.pickle', 'rb')
tokenizer = pickle.load(f)

# Load Trained Model

import tensorflow as tf
from tensorflow.keras.models import load_model

model = load_model(root + 'Saved/model_lstm.hdf5')

# Testing

from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.utils import to_categorical
from sklearn import preprocessing

max_length = 120
trunc_type='post'
padding_type='post'

sentences = [str(text) for text in df.text]
sequences = tokenizer.texts_to_sequences(sentences)
X_test = pad_sequences(sequences, maxlen=max_length, padding=padding_type, truncating=trunc_type)

label_encoder = preprocessing.LabelEncoder()
df['label']= label_encoder.fit_transform(df['label'])  
y_test = np.array(df['label'])
y_test = to_categorical(y_test)

label_encoder.classes_

# Predict

pred = model.predict(X_test)
y_pred = np.argmax(pred, axis=-1)

y_test = [np.argmax(y, axis=None, out=None) for y in y_test]

from sklearn.metrics import plot_confusion_matrix, confusion_matrix, classification_report, f1_score

print(classification_report(y_test, y_pred))

y = ''
while y != 'exit':
  y = input()
  ys = 
  pred = model.predict(y)
